seperates ci-search into seperate angular component with module with CUSTOM_ELEMENT_SCHEMA. Re-introduces login message for all geo search datasets.
