+++ src/app/components/header/info-bar/ci-search/ci-search.component.html
@@ -1,0 +1,1 @@
+<ci-search id="ASF_SEARCH"></ci-search>
