+++ src/app/components/header/info-bar/ci-search/index.ts
@@ -1,0 +1,2 @@
+export * from './ci-search.component';
+export * from './ci-search.module';
