+++ src/app/components/header/info-bar/info-bar.component.html
@@ -48,6 +48,6 @@
     </span>
   </div>
   <div fxFlex class="last-info">
-    <ci-search id="ASF_SEARCH"></ci-search>
+    <app-ci-search></app-ci-search>
   </div>
 </div>
