+++ src/app/components/results-menu/scene-detail/scene-detail.component.html
@@ -23,7 +23,7 @@
 
     <div class="detail-card-info">
       <p class="faint-text accessing-text">
-        <span *ngIf="!isLoggedIn && isRestrictedDataset()">
+        <span *ngIf="!isLoggedIn">
           Accessing this data requires you to log in.
         </span>
